Remove `include/grpcpp/impl/codegen/channel_interface.h`
