+++ CMakeLists.txt
@@ -3311,6 +3311,7 @@ foreach(_hdr
   include/grpcpp/impl/call_hook.h
   include/grpcpp/impl/call_op_set_interface.h
   include/grpcpp/impl/channel_argument_option.h
+  include/grpcpp/impl/channel_interface.h
   include/grpcpp/impl/client_unary_call.h
   include/grpcpp/impl/codegen/async_generic_service.h
   include/grpcpp/impl/codegen/async_stream.h
@@ -3982,6 +3983,7 @@ foreach(_hdr
   include/grpcpp/impl/call_hook.h
   include/grpcpp/impl/call_op_set_interface.h
   include/grpcpp/impl/channel_argument_option.h
+  include/grpcpp/impl/channel_interface.h
   include/grpcpp/impl/client_unary_call.h
   include/grpcpp/impl/codegen/async_generic_service.h
   include/grpcpp/impl/codegen/async_stream.h
