+++ include/grpcpp/impl/client_unary_call.h
@@ -20,8 +20,8 @@
 #define GRPCPP_IMPL_CLIENT_UNARY_CALL_H
 
 #include <grpcpp/impl/call.h>
+#include <grpcpp/impl/channel_interface.h>
 #include <grpcpp/impl/codegen/call_op_set.h>
-#include <grpcpp/impl/codegen/channel_interface.h>
 #include <grpcpp/impl/codegen/core_codegen_interface.h>
 #include <grpcpp/support/config.h>
 #include <grpcpp/support/status.h>
