+++ include/grpcpp/support/async_unary_call.h
@@ -23,8 +23,8 @@
 #include <grpcpp/client_context.h>
 #include <grpcpp/impl/call.h>
 #include <grpcpp/impl/call_op_set_interface.h>
+#include <grpcpp/impl/channel_interface.h>
 #include <grpcpp/impl/codegen/call_op_set.h>
-#include <grpcpp/impl/codegen/channel_interface.h>
 #include <grpcpp/impl/codegen/core_codegen_interface.h>
 #include <grpcpp/impl/service_type.h>
 #include <grpcpp/server_context.h>
