+++ include/grpcpp/support/client_callback.h
@@ -23,9 +23,9 @@
 #include <functional>
 
 #include <grpcpp/impl/call.h>
+#include <grpcpp/impl/channel_interface.h>
 #include <grpcpp/impl/codegen/call_op_set.h>
 #include <grpcpp/impl/codegen/callback_common.h>
-#include <grpcpp/impl/codegen/channel_interface.h>
 #include <grpcpp/impl/codegen/core_codegen_interface.h>
 #include <grpcpp/impl/codegen/sync.h>
 #include <grpcpp/support/config.h>
