+++ src/compiler/cpp_generator.cc
@@ -1650,7 +1650,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "functional",
         "grpcpp/support/async_stream.h",
         "grpcpp/support/async_unary_call.h",
-        "grpcpp/impl/codegen/channel_interface.h",
+        "grpcpp/impl/channel_interface.h",
         "grpcpp/impl/client_unary_call.h",
         "grpcpp/support/client_callback.h",
         "grpcpp/support/message_allocator.h",
