+++ tools/doxygen/Doxyfile.c++
@@ -958,6 +958,7 @@ include/grpcpp/impl/call.h \
 include/grpcpp/impl/call_hook.h \
 include/grpcpp/impl/call_op_set_interface.h \
 include/grpcpp/impl/channel_argument_option.h \
+include/grpcpp/impl/channel_interface.h \
 include/grpcpp/impl/client_unary_call.h \
 include/grpcpp/impl/codegen/async_generic_service.h \
 include/grpcpp/impl/codegen/async_stream.h \
